+++ middleware/connect/middleware.go
@@ -6,18 +6,23 @@ import (
 
 	"github.com/bufbuild/connect-go"
 	"github.com/streamingfast/dauth"
+	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-func NewAuthInterceptor(check dauth.Authenticator) *AuthInterceptor {
+var _ connect.Interceptor = (*AuthInterceptor)(nil)
+
+func NewAuthInterceptor(check dauth.Authenticator, logger *zap.Logger) *AuthInterceptor {
 	return &AuthInterceptor{
-		check: check,
+		check:  check,
+		logger: logger,
 	}
 }
 
 type AuthInterceptor struct {
-	check dauth.Authenticator
+	check  dauth.Authenticator
+	logger *zap.Logger
 }
 
 // WrapUnary implements [Interceptor] by applying the interceptor function.
@@ -30,7 +35,7 @@ func (i *AuthInterceptor) WrapUnary(next connect.UnaryFunc) connect.UnaryFunc {
 
 		childCtx, err := validateAuth(ctx, path, headers, peerAddr, i.check)
 		if err != nil {
-			return nil, obfuscateErrorMessage(err)
+			return nil, obfuscateErrorMessage(err, i.logger)
 		}
 
 		return next(childCtx, req)
@@ -46,7 +51,7 @@ func (i *AuthInterceptor) WrapStreamingHandler(next connect.StreamingHandlerFunc
 
 		childCtx, err := validateAuth(ctx, path, headers, peerAddr, i.check)
 		if err != nil {
-			return obfuscateErrorMessage(err)
+			return obfuscateErrorMessage(err, i.logger)
 		}
 
 		return next(childCtx, conn)
@@ -58,14 +63,18 @@ func (i *AuthInterceptor) WrapStreamingClient(next connect.StreamingClientFunc) 
 	return next
 }
 
-func obfuscateErrorMessage(err error) error {
+func obfuscateErrorMessage(err error, logger *zap.Logger) error {
 	if st, ok := status.FromError(err); ok {
 		msg := st.Message()
 		switch st.Code() {
 		case codes.Internal, codes.Unavailable, codes.Unknown:
+			logger.Error("authentication service via Connect-Web middleware fatal error", zap.Error(err))
 			msg = "error with authentication service, please try again later"
 		}
 		return connect.NewError(connect.Code(st.Code()), errors.New(msg))
+	} else {
+		logger.Error("authentication service via Connect-Web middleware non-gRPC error", zap.Error(err))
 	}
+
 	return err
 }